+++ CHANGES.rst
@@ -51,6 +51,7 @@ Changelog
    Commit items for NEXT VERSION
    ..............................
 
+   - feat(register_config): separate extract and get registry into an optional two step process
    - test(test_registry_config): unlink file used by multiple processes suppress OSError
    - chore(setup.cfg): remove file
    - ci(quality): target doc use py310